+++ HappyDNS/include/HappyDNS/QNDes.h
@@ -1,0 +1,1 @@
+../../Util/QNDes.h