+++ HappyDNS/include/HappyDNS/QNDnsMessage.h
@@ -1,0 +1,1 @@
+../../Dns/QNDnsMessage.h