+++ HappyDNS/include/HappyDNS/QNDnsUdpResolver.h
@@ -1,0 +1,1 @@
+../../Dns/QNDnsUdpResolver.h