+++ HappyDNS/include/HappyDNS/QNResolver.h
@@ -1,0 +1,1 @@
+../../Local/QNResolver.h