+++ HappyDNS/include/HappyDNS/QNResolverDelegate.h
@@ -1,0 +1,1 @@
+../../Common/QNResolverDelegate.h